Move ProgDialog to interfaces sub-module.
